+++ src/features/board/boardSlice.ts
@@ -1,220 +1,22 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit"
-import { ISquare, SquareColor } from "../square/squareSlice"
+import { SquareColor } from "../square/squareSlice"
 import { AppThunk, RootState } from "../../app/store"
 import {
-  canIMove,
-  canIMoveOrBeat,
   dragPiece,
   dropPiece,
-  getCoordFromPosition,
-  getPositionFromCoords,
-  IPiece,
-  isSquareCanBeBeaten,
+  modifyPieceType,
   movePieceFromTo,
   PieceColor,
-  PiecePosition,
   PieceType,
   placePiece
 } from "../piece/pieceSlice"
 import { checkTo, clearCheck, draw, mateTo } from "../game/gameSlice"
-import { traverseInTime } from "../history/historySlice";
-
-export interface Board {
-  squares: ISquare[][],
-  activeSquare: string,
-  possibleMovements: { [key: PiecePosition]: null },
-
-  /**
-   * TODO: Не давать ходить, если подставляемся под шах или мат
-   */
-
-  /**
-   * TODO: Castling
-   *
-   */
-  // Рокировка невозможна:
-  //   если король по ходу партии уже делал ходы (включая ход-рокировку);
-  //   если ладья походила, то с ней уже невозможна рокировка
-  // Рокировка временно невозможна:
-  //   если король под шахом. В результате рокировки пройдёт через битое поле или встанет под шах[2].
-  //   если между королём и ладьей, предназначенными для рокировки, находится другая фигура (своя или чужая)[3].
-}
-
-/**
- * Find square by position
- * @param position
- * @param squares
- */
-const findSquare = (position: PiecePosition, squares: ISquare[][]): ISquare => {
-  for (let i = 0; i < squares.length; i++) {
-    for (let j = 0; j < squares[0].length; j++) {
-      if (squares[i][j].position === position) return squares[i][j]
-    }
-  }
-
-  return {} as ISquare
-}
-
-/**
- * Finds king by color
- * @param color
- * @param squares
- */
-const findKingsSquareByColor = (color: PieceColor, squares: ISquare[][]): ISquare => {
-  for (let i = 0; i < squares.length; i++) {
-    for (let j = 0; j < squares[0].length; j++) {
-      if (squares[i][j].piece?.type === PieceType.KING
-        && squares[i][j].piece?.color === color
-      ) {
-        return squares[i][j]
-      }
-    }
-  }
-
-  return {} as ISquare
-}
-
-/**
- * Checks if king has place to run
- * @param kingSquare
- * @param squares
- */
-const kingCanEscape = (kingSquare: ISquare, squares: ISquare[][]): boolean => {
-  const [ y, x ] = kingSquare.coords
-  const positions = [
-    [ y - 1, x - 1 ],
-    [ y - 1, x ],
-    [ y - 1, x + 1 ],
-    [ y, x - 1 ],
-    [ y, x + 1 ],
-    [ y + 1, x - 1 ],
-    [ y + 1, x ],
-    [ y + 1, x + 1 ],
-  ].filter(([ y, x ]) => !!squares[y] && squares[y][x])
-
-  return positions
-    .some(([ y, x ]) => {
-      return canIMoveOrBeat(kingSquare.piece!, squares[y][x].position, squares)
-    })
-}
-
-/**
- * Checks if there are obstacles between two squares
- * @param y0
- * @param x0
- * @param y1
- * @param x1
- * @param squares
- */
-export const haveObstaclesBetween = (y0: number, x0: number, y1: number, x1: number, squares: ISquare[][]): boolean => {
-  if (y0 === y1 && x0 === x1) return false
-
-  // If it's horizontal move
-  if (y0 === y1) {
-    // Check all horizontal pieces in between start and end
-    for (let i = Math.min(x0, x1) + 1; i < Math.max(x0, x1); i++) {
-      if (!!squares[y0][i]?.piece?.type) return true
-    }
-  }
-
-  // If it's vertical move
-  if (x0 === x1) {
-    // Check all vertical pieces in between start and end
-    for (let i = Math.min(y0, y1) + 1; i < Math.max(y0, y1); i++) {
-      if (!!squares[i][x0]?.piece?.type) return true
-    }
-  }
-
-  // If it's diagonal move
-  if (y1 !== y0 && x1 !== x0) {
-    // Check north-west or south-east
-    if ((x1 < x0 && y1 < y0) || (x1 > x0 && y1 > y0)) {
-      for (let i = Math.min(y0, y1) + 1, j = Math.min(x0, x1) + 1; i < Math.max(y0, y1) && j < Math.max(x0, x1); i++, j++) {
-        if (!!squares[i][j]?.piece?.type) return true
-      }
-    }
-
-    // Check south-west and north-east
-    if ((x1 < x0 && y1 > y0) || (x1 > x0 && y1 < y0)) {
-      for (let i = Math.max(y0, y1) - 1, j = Math.min(x0, x1) + 1; i > Math.min(y0, y1) && j < Math.max(x0, x1); i--, j++) {
-        if (!!squares[i][j]?.piece?.type) return true
-      }
-    }
-  }
-
-  return false
-}
-
-/**
- * Gets allied pieces
- * @param allyColor
- * @param squares
- */
-const getAlliedPieces = (allyColor: PieceColor, squares: ISquare[][]): IPiece[] => {
-  const alliedPieces = []
-
-  // Get allied pieces
-  for (let i = 0; i < squares.length; i++) {
-    for (let j = 0; j < squares[0].length; j++) {
-      if (!!squares[i][j].piece && squares[i][j].piece?.color === allyColor) {
-        alliedPieces.push(squares[i][j].piece!)
-      }
-    }
-  }
-
-  return alliedPieces
-}
-
-/**
- * Checks if someone can protect king
- */
-const someoneCanProtectKing = (kingSquare: ISquare, from: ISquare, squares: ISquare[][]): boolean => {
-  // Someone can go to any cell between king and threat
-  const [ y1, x1 ] = getCoordFromPosition(kingSquare.position)
-  const [ y0, x0 ] = getCoordFromPosition(from.position)
-  const alliedPieces = getAlliedPieces(kingSquare.piece!.color, squares)
-
-  // No one can protect from pawn or knight
-  if ([ PieceType.KNIGHT, PieceType.PAWN ].includes(from.piece?.type!)) return false
-
-  // No one can protect from contacting piece
-  if (Math.abs(y0 - y1) === 1 || Math.abs(x0 - x1) === 1) return false
-
-  // West/east
-  if (y0 === y1) {
-    for (let i = Math.min(x0, x1) + 1; i < Math.max(x0, x1); i++) {
-      if (alliedPieces.some((piece) => canIMove(piece, getPositionFromCoords(y0, i), squares))) return true
-    }
-  }
-
-  // North/south
-  if (x0 === x1) {
-    // Check all vertical pieces in between start and end
-    for (let i = Math.min(y0, y1) + 1; i < Math.max(y0, y1); i++) {
-      if (alliedPieces.some((piece) => canIMove(piece, getPositionFromCoords(i, x0), squares))) return true
-    }
-  }
-
-  // Diagonals
-  if (y1 !== y0 && x1 !== x0) {
-    // Check north-west or south-east
-    if ((x1 < x0 && y1 < y0) || (x1 > x0 && y1 > y0)) {
-      for (let i = Math.min(y0, y1) + 1, j = Math.min(x0, x1) + 1; i < Math.max(y0, y1) && j < Math.max(x0, x1); i++, j++) {
-        if (alliedPieces.some((piece) => canIMove(piece, getPositionFromCoords(i, j), squares))) return true
-      }
-    }
-
-    // Check south-west and north-east
-    if ((x1 < x0 && y1 > y0) || (x1 > x0 && y1 < y0)) {
-      // console.log(alliedPieces)
-      for (let i = Math.max(y0, y1) - 1, j = Math.min(x0, x1) + 1; i > Math.min(y0, y1) && j < Math.max(x0, x1); i--, j++) {
-        if (alliedPieces.some((piece) => canIMove(piece, getPositionFromCoords(i, j), squares))) return true
-      }
-    }
-  }
-
-  return false
-}
+import { traverseInTime } from "../history/historySlice"
+import { IPiece, PiecePosition } from "../piece/types"
+import { ISquare } from "../square/types"
+import { Board } from "./types"
+import { findKingsSquareByColor, findSquare, getAlliedPieces, kingCanEscape, someoneCanProtectKing } from "./utils"
+import { canIMoveOrBeat, getCoordFromPosition, isSquareCanBeBeaten } from "../piece/utils"
 
 /**
  * Processes check/mate situation
@@ -271,107 +73,130 @@ export const processGameState = (): AppThunk => (dispatch, getState) => {
   }
 }
 
+/**
+ * TODO: Не давать ходить, если подставляемся под шах или мат
+ */
+
+/**
+ * TODO: Castling
+ *
+ */
+  // Рокировка невозможна:
+  //   если король по ходу партии уже делал ходы (включая ход-рокировку);
+  //   если ладья походила, то с ней уже невозможна рокировка
+  // Рокировка временно невозможна:
+  //   если король под шахом. В результате рокировки пройдёт через битое поле или встанет под шах[2].
+  //   если между королём и ладьей, предназначенными для рокировки, находится другая фигура (своя или чужая)[3].
+
+
 const boardSlice = createSlice({
-  name: "field",
-  initialState: {
-    squares: [],
-    activeSquare: "",
-    possibleMovements: {}
-  } as Board,
-  reducers: {
-    initSquares: (state: Board) => {
-      const squares =
-        new Array(8)
-          .fill(null)
-          .map(() =>
-            new Array(8)
-              .fill(null)
-              .map(() => ({} as ISquare)))
+    name: "field",
+    initialState: {
+      squares: [],
+      activeSquare: "",
+      possibleMovements: {}
+    } as Board,
+    reducers: {
+      initSquares: (state: Board) => {
+        const squares =
+          new Array(8)
+            .fill(null)
+            .map(() =>
+              new Array(8)
+                .fill(null)
+                .map(() => ({} as ISquare)))
 
-      const n = squares.length
-      const m = squares[0].length
+        const n = squares.length
+        const m = squares[0].length
 
-      for (let i = 0, rowCount = 8; i < n; i++, rowCount--) {
-        const isEvenRow = i % 2 === 0
+        for (let i = 0, rowCount = 8; i < n; i++, rowCount--) {
+          const isEvenRow = i % 2 === 0
 
-        for (let j = 0, charCode = 97; j < m; j++, charCode++) {
-          const name = String.fromCharCode(charCode).toUpperCase() + rowCount
-          if (!isEvenRow) {
-            squares[i][j].color = j % 2 === 0 ? SquareColor.BLACK : SquareColor.WHITE
-          } else {
-            squares[i][j].color = j % 2 === 0 ? SquareColor.WHITE : SquareColor.BLACK
-          }
+          for (let j = 0, charCode = 97; j < m; j++, charCode++) {
+            const name = String.fromCharCode(charCode).toUpperCase() + rowCount
+            if (!isEvenRow) {
+              squares[i][j].color = j % 2 === 0 ? SquareColor.BLACK : SquareColor.WHITE
+            } else {
+              squares[i][j].color = j % 2 === 0 ? SquareColor.WHITE : SquareColor.BLACK
+            }
 
-          // Set square's params
-          squares[i][j] = {
-            ...squares[i][j],
-            position: name,
-            coords: [ i, j ]
-          }
-        }
-      }
-
-      state.squares = squares
-    },
-  }, extraReducers: (builder) => {
-    builder.addCase(placePiece, (state, action) => {
-      const { position, type, color } = action.payload
-      const [ rank, file ] = getCoordFromPosition(position)
-
-      state.squares[rank][file].piece = {
-        type,
-        color,
-        position: position.toUpperCase(),
-        moved: false
-      }
-    })
-    builder
-      .addCase(dragPiece, (state, action: PayloadAction<IPiece>) => {
-        const { squares } = state
-        const { position } = action.payload
-        const square = findSquare(position, squares)
-
-        // If piece is no longer at the start position
-        if (!square.piece) return
-
-        state.activeSquare = position
-
-        for (let i = 0; i < squares.length; i++) {
-          for (let j = 0; j < squares[0].length; j++) {
-            if (canIMoveOrBeat(square.piece!, squares[i][j].position, squares)) {
-              state.possibleMovements[squares[i][j].position] = null
+            // Set square's params
+            squares[i][j] = {
+              ...squares[i][j],
+              position: name,
+              coords: [ i, j ]
             }
           }
         }
+
+        state.squares = squares
+      },
+    }, extraReducers: (builder) => {
+      builder.addCase(placePiece, (state, action) => {
+        const { position, type, color } = action.payload
+        const [ rank, file ] = getCoordFromPosition(position)
+
+        state.squares[rank][file].piece = {
+          type,
+          color,
+          position: position.toUpperCase(),
+          moved: false
+        }
       })
-    // Clear active square and possible movements
-    builder.addCase(dropPiece, (state) => {
-      state.activeSquare = ""
-      state.possibleMovements = {}
-    })
-    // Make move
-    builder.addCase(movePieceFromTo, (state, action) => {
+      builder.addCase(modifyPieceType, (state, action) => {
+        const { newType, piece: { position } } = action.payload
+        const [ rank, file ] = getCoordFromPosition(position)
+
+        state.squares[rank][file].piece!.type = newType
+      })
+      builder
+        .addCase(dragPiece, (state: Board, action: PayloadAction<IPiece>) => {
+          const { squares } = state
+          const { position } = action.payload
+          const square = findSquare(position, squares)
+
+          // If piece is no longer at the start position
+          if (!square.piece) return
+
+          state.activeSquare = position
+
+          for (let i = 0; i < squares.length; i++) {
+            for (let j = 0; j < squares[0].length; j++) {
+              if (canIMoveOrBeat(square.piece!, squares[i][j].position, squares)) {
+                const position: PiecePosition = squares[i][j].position
+                state.possibleMovements[position] = null
+              }
+            }
+          }
+        })
       // Clear active square and possible movements
-      state.possibleMovements = {}
-      state.activeSquare = ""
+      builder.addCase(dropPiece, (state) => {
+        state.activeSquare = ""
+        state.possibleMovements = {}
+      })
+      // Make move
+      builder.addCase(movePieceFromTo, (state, action) => {
+        // Clear active square and possible movements
+        state.possibleMovements = {}
+        state.activeSquare = ""
 
-      const { from, to, piece } = action.payload
+        const { from, to, piece } = action.payload
 
-      const [ rankFrom, fileFrom ] = getCoordFromPosition(from)
-      const [ rankTo, fileTo ] = getCoordFromPosition(to)
+        const [ rankFrom, fileFrom ] = getCoordFromPosition(from)
+        const [ rankTo, fileTo ] = getCoordFromPosition(to)
 
-      delete state.squares[rankFrom][fileFrom].piece
+        delete state.squares[rankFrom][fileFrom].piece
 
-      state.squares[rankTo][fileTo].piece = {
-        ...piece,
-        moved: true
-      }
-    })
-    builder.addCase(traverseInTime, (state, action) => {
-      return action.payload.board
-    })
-  }
-})
+        state.squares[rankTo][fileTo].piece = {
+          ...piece,
+          moved: true
+        }
+      })
+      builder.addCase(traverseInTime, (state, action) => {
+        return action.payload.board
+      })
+    }
+  })
 
 export const initPieces =
   (): AppThunk =>