+++ src/features/board/types.d.ts
@@ -1,0 +1,8 @@
+import { PiecePosition } from "../piece/types"
+import { ISquare } from "../square/types"
+
+export interface Board {
+  squares: ISquare[][],
+  activeSquare: string,
+  possibleMovements: { [key: PiecePosition]: null },
+}