+++ src/features/game/types.d.ts
@@ -1,0 +1,7 @@
+export interface Game {
+  turn: PieceColor,
+  checkTo: PieceColor | null,
+  mateTo: PieceColor | null,
+  draw: boolean,
+  gameOver: boolean,
+}