+++ src/features/history/types.d.ts
@@ -1,0 +1,6 @@
+export type HistoryItem = {
+  name: string,
+  board: Board,
+  game: Game,
+  piece: { current: IPiece }
+}