+++ src/features/square/Square.tsx
@@ -1,26 +1,28 @@
-import styles from './Square.module.css'
-import { ISquare, SquareColor } from "./squareSlice";
+import styles from "./Square.module.css"
+import { SquareColor } from "./squareSlice"
 import { FC } from "react"
-import classNames from "classnames";
-import { Piece } from "../piece/Piece";
-import { useDrop } from 'react-dnd'
+import classNames from "classnames"
+import { Piece } from "../piece/Piece"
+import { useDrop } from "react-dnd"
 import { useAppDispatch, useAppSelector } from "../../app/hooks"
 import { selectActiveSquare, selectPossibleMovements, selectSquares } from "../board/boardSlice"
-import { selectCurrentPiece, canIMoveOrBeat, movePieceTo } from "../piece/pieceSlice"
+import { movePieceTo, selectCurrentPiece } from "../piece/pieceSlice"
+import { ISquare } from "./types"
+import { canIMoveOrBeat } from "../piece/utils"
 
 interface Props {
   square: ISquare
 }
 
 export const Square: FC<Props> = ({ square }) => {
-  const dispatch = useAppDispatch();
+  const dispatch = useAppDispatch()
   const currentPiece = useAppSelector(selectCurrentPiece)
   const squares = useAppSelector(selectSquares)
   const activeSquare = useAppSelector(selectActiveSquare)
   const possibleMovements = useAppSelector(selectPossibleMovements)
 
   const [ , drop ] = useDrop(() => ({
-    accept: 'piece',
+    accept: "piece",
     drop: () => {
       dispatch(movePieceTo(square.position))
     },
@@ -33,13 +35,13 @@ export const Square: FC<Props> = ({ square }) => {
   }), [ currentPiece, square.position ])
 
   return <div
-    ref={ drop }
-    className={ classNames([
+    ref={drop}
+    className={classNames([
       styles.Square,
       square.position === activeSquare ? styles.active : null,
       square.position in possibleMovements ? styles.possible : null,
       square.color === SquareColor.BLACK
         ? styles.black
-        : styles.white ]) }
-  >{ square.piece ? <Piece square={ square }/> : null }</div>
+        : styles.white ])}
+  >{square.piece ? <Piece square={square}/> : null}</div>
 }