+++ src/features/square/types.d.ts
@@ -1,0 +1,7 @@
+export interface ISquare {
+  color: SquareColor,
+  piece?: IPiece,
+  position: string,
+  active: boolean,
+  coords: [ y: number, x: number ]
+}